refactor working well, all tested

+++ desktop-apps/project/menu.py
@@ -1,20 +1,17 @@
-
-from PyQt5.QtGui import QCursor, QKeySequence
+from PyQt5.QtGui import QKeySequence
 from PyQt5.QtWidgets import QMenu, QAction, QMessageBox, QFileDialog
 
 from messages import Message
 
-class Menu(QMenu):
 
+class Menu(QMenu):
     def __init__(self, main, explorer, editor, model):
         super().__init__()
         self.explorer = explorer
         self.editor = editor
         self.model = model
         self.main = main
-        self.message = Message()
-
-        self.menu_actions()
+        self.message = Message(main)
 
     def menu_actions(self):
 
@@ -26,7 +23,7 @@ class Menu(QMenu):
         self.open_action = QAction("&Open file...")
         self.open_action.triggered.connect(self.show_open_dialog)
         self.open_action.setShortcut(QKeySequence.Open)
-        self.main.file_menu.addAction(self.open_action)       
+        self.main.file_menu.addAction(self.open_action)
 
         self.save_action = QAction("&Save")
         self.save_action.triggered.connect(self.save)
@@ -42,35 +39,34 @@ class Menu(QMenu):
         self.about_action.triggered.connect(self.message.show_about_dialog)
         self.main.help_menu.addAction(self.about_action)
 
-
     def new_document(self):
 
-        ''' This function creates a new document and saves if another one is open
-            and have changes '''
+        """ This function creates a new document and saves if another one is open
+            and have changes """
 
-        self.safe_close()
+        if self.safe_close() == "cancel":
+            return
         self.editor.clear()
         self.main.file_path = None
 
     def save(self):
 
-        ''' Standard save function. Will ask the file_path throught a dialog
-        if there is no file_path '''
+        """ Standard save function. Will ask the file_path throught a dialog
+        if there is no file_path """
 
         if self.main.file_path is None:
             return self.show_save_dialog()
         else:
-            with open(self.main.file_path, 'w') as f:
+            with open(self.main.file_path, "w") as f:
                 f.write(self.editor.toPlainText())
             self.editor.document().setModified(False)
             return True
 
-
     def show_save_dialog(self):
 
-        ''' Show the save dialog is there is no file_path'''
+        """ Show the save dialog is there is no file_path """
 
-        self.filename, _ = QFileDialog.getSaveFileName(self, 'Save as...')
+        self.filename, _ = QFileDialog.getSaveFileName(self.main, "Save as...")
         if self.filename:
             self.main.file_path = self.filename
             self.save()
@@ -79,24 +75,23 @@ class Menu(QMenu):
 
     def safe_close(self):
 
-        ''' If the document is modified, call message.ask_for_confirmation '''
+        """ If the document is modified, call message.ask_for_confirmation """
 
         if self.editor.document().isModified():
             answer = self.message.ask_for_confirmation()
             if answer == QMessageBox.Save:
-                if not self.save():
-                    return
+                self.save()
             elif answer == QMessageBox.Cancel:
-                return
+                return "cancel"
 
     def show_open_dialog(self):
 
-        ''' Function for open files saved on your device '''
+        """ Function for open files saved on your device """
 
-        self.filename, _ = QFileDialog.getOpenFileName(self, 'Open...')
+        self.filename, _ = QFileDialog.getOpenFileName(self.main, "Open...")
         if self.filename:
             file_contents = ""
-            with open(self.filename, 'r') as f:
+            with open(self.filename, "r") as f:
                 file_contents = f.read()
             self.editor.setPlainText(file_contents)
             self.main.file_path = self.filename
